welaser/testScripts/requirements.txt
kafka-python==2.0.2
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
paho-mqtt==1.6.1
